// kitchen-blog/src/app/model/article.ts
export class Article {
    id: number
    title: string;
    desc: string;
    userId: number;
    username: string;
    difficulty: string;
    ingredients: string[];
    preparationTime: number;
    cockingTime: number;
    preparation: string[];
    type: string;
    etapes: string[];
    constructor() {
    }
}